divide the components and the page with the logic and the element form
