+++ src/App.js
@@ -9,7 +9,7 @@ from 'react-router-dom'
 import Header from './components/Header'
 import { Footer } from './components/Footer'
 import Home from './pages/Home'
-import FormSignUp from './pages/FormSignUp'
+import SignUp from './pages/SignUp'
 import FormSignIn from './pages/FormSignIn'
 import SanitaryRegister from './pages/SanitaryRegister'
 import ClientProfile from './pages/ClientProfile'
@@ -25,7 +25,7 @@ function App () {
     <Header></Header>
       <Switch>
         <Route exact path="/" component={Home}/>
-        <Route exact path="/sign-up" component={FormSignUp}/>
+        <Route exact path="/sign-up" component={SignUp}/>
         <Route exact path="/sign-in" component={FormSignIn}/>
         <Route exact path="/sanitary-register" component={SanitaryRegister}/>
         <Route exact path="/client-profile" component={ClientProfile}/>
